apps/backend/pyproject.toml
[tool.poetry]
name = "wataomi-backend"
version = "1.0.0"
description = "WataOmi Backend API"
authors = ["WataOmi Team"]

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.109.0"
uvicorn = {extras = ["standard"], version = "^0.27.0"}
sqlmodel = "^0.0.14"
psycopg2-binary = "^2.9.9"
asyncpg = "^0.29.0"
redis = "^5.0.1"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.6"
pydantic = {extras = ["email"], version = "^2.5.3"}
pydantic-settings = "^2.1.0"
httpx = "^0.26.0"
supabase = "^2.3.0"
google-generativeai = "^0.3.2"
qdrant-client = "^1.7.0"
cloudinary = "^1.38.0"
casdoor-python-sdk = "^1.3.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.4"
black = "^23.12.1"
mypy = "^1.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
